requirements.txt
Flask==2.0.2
requests==2.26.0
openai==0.10.2
pandas==1.4.0
gspread==5.1.1
oauth2client==4.1.3
python-dotenv==0.19.1